backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
